fix codestyle and put qsort in separate module
